+++ sw.js
@@ -13,7 +13,7 @@
  */
 'use strict';
 
-var VERSION = 'v8';
+var VERSION = 'v9';
 var RUNTIME_CACHE_VERSION = 'v1';
 var SHELL_CACHE = 'stormscope-shell-' + VERSION;
 var TILE_CACHE = 'stormscope-tiles-' + RUNTIME_CACHE_VERSION;
@@ -39,6 +39,7 @@ var SHELL_ASSETS = [
   './js/camera-store.js',
   './js/saved-state.js',
   './js/multi-camera.js',
+  './js/context-layers.js',
   './js/app.js',
   './vendor/leaflet/leaflet.css',
   './vendor/leaflet/leaflet.js',